Rename Multiplex to Multicast
Appearantly, multiplex is N to 1 while this thing is 1 to N. Yigit should consider checking what things mean before using them in code ¯\_(ツ)_/¯
